Code/backend/ParkMindfulness/requirements.txt
asgiref==3.6.0
colorama==0.4.6
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.13.0
djangorestframework-yaml==2.0.0
Pillow==9.4.0
PyJWT==2.6.0
pypng==0.20220715.0
pytz==2022.7.1
qrcode==7.4.2
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2022.7
gunicorn==20.1.0
whitenoise==6.4.0
